разложил по пакетам
