+++ app/src/main/java/com/example/popularlibs_homrworks/model/repositories/repo/RetrofitGithubRepositoriesRepo.kt
@@ -4,8 +4,8 @@ import com.example.popularlibs_homrworks.model.api.IDataSource
 import com.example.popularlibs_homrworks.model.entity.GithubUser
 import com.example.popularlibs_homrworks.model.entity.GithubUserRepos
 import com.example.popularlibs_homrworks.model.repositories.repo.cash.IRoomRepositiriesRepoCash
-import com.example.popularlibs_homrworks.model.room.database.Database
-import com.example.popularlibs_homrworks.model.room.network.INetworkStatus
+import com.example.popularlibs_homrworks.model.room.Database
+import com.example.popularlibs_homrworks.model.network.INetworkStatus
 import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.schedulers.Schedulers
