+++ app/src/main/java/com/example/popularlibs_homrworks/model/repositories/repo/cash/IRoomRepositiriesRepoCash.kt
@@ -2,7 +2,7 @@ package com.example.popularlibs_homrworks.model.repositories.repo.cash
 
 import com.example.popularlibs_homrworks.model.entity.GithubUser
 import com.example.popularlibs_homrworks.model.entity.GithubUserRepos
-import com.example.popularlibs_homrworks.model.room.database.Database
+import com.example.popularlibs_homrworks.model.room.Database
 import io.reactivex.rxjava3.core.Single
 
 //интерфейс для кэширования списка репозиториев пользователя user из сети в базу данных
