+++ app/src/main/java/com/example/popularlibs_homrworks/model/repositories/users/RetrofitGithubUsersRepo.kt
@@ -3,8 +3,8 @@ package com.example.popularlibs_homrworks.model.repositories.users
 import com.example.popularlibs_homrworks.model.api.IDataSource
 import com.example.popularlibs_homrworks.model.entity.GithubUser
 import com.example.popularlibs_homrworks.model.repositories.users.cash.IRoomGithubUsersCache
-import com.example.popularlibs_homrworks.model.room.database.Database
-import com.example.popularlibs_homrworks.model.room.network.INetworkStatus
+import com.example.popularlibs_homrworks.model.room.Database
+import com.example.popularlibs_homrworks.model.network.INetworkStatus
 import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.schedulers.Schedulers
 
@@ -13,8 +13,8 @@ import io.reactivex.rxjava3.schedulers.Schedulers
 // Таким образом, мы позволяем репозиторию самостоятельно следить за тем, чтобы сетевые вызовы
 // выполнялись именно в io-потоке. Всегда лучше поступать именно таким образом, даже когда речь
 // не идёт о сети — во избежание выполнения операций в неверном потоке в вызывающем коде.
-class RetrofitGithubUsersRepo(val api:IDataSource, val networkStatus:INetworkStatus,
- val db: Database, val roomCash: IRoomGithubUsersCache
+class RetrofitGithubUsersRepo(val api:IDataSource, val networkStatus: INetworkStatus,
+                              val db: Database, val roomCash: IRoomGithubUsersCache
 ):IGithubUsersRepo {
     //метод  интерфейса IGithubUsersRepo getUsers() - в зависимости от статуса сети
     //мы или получаем данные из сети, записывая их в базу данных с помощью Room через map
