+++ app/src/main/java/com/example/popularlibs_homrworks/model/repositories/users/cash/IRoomGithubUsersCache.kt
@@ -1,7 +1,7 @@
 package com.example.popularlibs_homrworks.model.repositories.users.cash
 
 import com.example.popularlibs_homrworks.model.entity.GithubUser
-import com.example.popularlibs_homrworks.model.room.database.Database
+import com.example.popularlibs_homrworks.model.room.Database
 import io.reactivex.rxjava3.core.Single
 
 //интерфейс для кэширования списка пользователей из сети в базу данных
