+++ app/src/main/java/com/example/popularlibs_homrworks/model/repositories/users/cash/RoomGithubUsersCache.kt
@@ -1,8 +1,8 @@
 package com.example.popularlibs_homrworks.model.repositories.users.cash
 
 import com.example.popularlibs_homrworks.model.entity.GithubUser
-import com.example.popularlibs_homrworks.model.room.database.Database
-import com.example.popularlibs_homrworks.model.room.tablesroom.RoomGithubUser
+import com.example.popularlibs_homrworks.model.room.Database
+import com.example.popularlibs_homrworks.model.room.tables.RoomGithubUser
 import io.reactivex.rxjava3.core.Single
 
 //класс для реализации кэширования списка пользователей из сети в базу данных
@@ -14,8 +14,10 @@ class RoomGithubUsersCache:    IRoomGithubUsersCache {
        return Single.fromCallable{ //создаём  Single из списка, по пути пишем в базу
             // map для базы, так как классы разные
             val roomUsers = githubUserList.map {user->
-                RoomGithubUser(user.id?:"", user.login?:"",
-                    user.avatarUrl?:"", user.repos_url?:"" )
+                RoomGithubUser(
+                    user.id ?: "", user.login ?: "",
+                    user.avatarUrl ?: "", user.repos_url ?: ""
+                )
             }
             db.userDao.insert(roomUsers) //пишем в базу
             return@fromCallable  githubUserList //возвращаем users  в виде Single<List<GithubUserRepos>>
