+++ app/src/main/java/com/example/popularlibs_homrworks/view/fragments/user/UserFragment.kt
@@ -12,8 +12,8 @@ import com.example.popularlibs_homrworks.model.api.ApiHolder
 import com.example.popularlibs_homrworks.model.entity.GithubUser
 import com.example.popularlibs_homrworks.model.repositories.repo.RetrofitGithubRepositoriesRepo
 import com.example.popularlibs_homrworks.model.repositories.repo.cash.RoomRepositoriesRepoCash
-import com.example.popularlibs_homrworks.model.room.database.Database
-import com.example.popularlibs_homrworks.model.room.network.AndroidNetworkStatus
+import com.example.popularlibs_homrworks.model.room.Database
+import com.example.popularlibs_homrworks.model.network.AndroidNetworkStatus
 import com.example.popularlibs_homrworks.presenters.user.UserRepoPresenter
 import com.example.popularlibs_homrworks.view.adapters.user.UserRepoAdapter
 import com.example.popularlibs_homrworks.view.fragments.BackButtonListener
@@ -33,7 +33,10 @@ class UserFragment(val user: GithubUser) : MvpAppCompatFragment(),
         UserRepoPresenter(
             AndroidSchedulers.mainThread(),
             RetrofitGithubRepositoriesRepo(
-                ApiHolder.api, AndroidNetworkStatus(App.instance),
+                ApiHolder.api,
+                AndroidNetworkStatus(
+                    App.instance
+                ),
                 Database.getInstance(), RoomRepositoriesRepoCash()
             ),
             App.instance.router,
